requirements.txt
alembic==0.9.6
bottle==0.12.13
bottle-sqlalchemy==0.4.3
click==6.7
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.15
